dbmw: 修复 affected-rows 误判导致的 7 处 insert 语义 bug
Ghidra 逐函数核对发现此前把 ORIG 的 64 位 affected 检查（getAffectedRowCount
返回 my_ulonglong，调用点 or eax,edx 判非零）误判为「or %edx 死代码恒真」，
导致 7 处 insert 分支语义错位（真实 bug）：
- SaveGuildSkill：insert(0x4e59) 完全缺失（新公会永远不写 guild_skill）
- SavePowerWarPoint / OnSavePowerWarStatueRanker / ChangeGuildNotifyMessage /
  QueryHWspecCreate(3 分支)：insert 仅在 exec 失败时执行，ORIG 为
  exec!=1 || affected==0（新 occ_time/新记录不入库）
- OnGoldcardEventStatistic：循环条件 && 应为 ||（create/open 任一非零），
  insert(0x4f02) 缺失
- OnServerMatchData：insert 时机完全反了（ORIG 是 exec 成功且 affected==0
  时 insert；原实现 exec 失败才 insert）
- OnSavePowerWarStatueRanker：补 exec(0x4ecc) 删除执行（原来只 set_query
  不执行）；update 的 set_query 返回值按 ORIG 忽略
getAffectedRowCount 声明/实现改 unsigned long long（CMySql 9/9 IDENTICAL）。
全量水位 609/513/1443 无回归。exec 返回类型 int→bool 的全局码型对齐
（240 处调用点改 != 1）留待下一批。

+++ dnf_decompile/source/dbmw/CMySql.cpp
@@ -99,9 +99,9 @@ char CMySql::set_reconnect_option()
     return 1;
 }
 
-int CMySql::getAffectedRowCount()
+unsigned long long CMySql::getAffectedRowCount()
 {
-    return (int)mysql_affected_rows(m_mysql);
+    return mysql_affected_rows(m_mysql);
 }
 
 int CMySql::get_int(int col, unsigned int& v)
